+++ src/main/java/org/ddolib/ddo/core/Solver.java
@@ -13,7 +13,7 @@ public interface Solver {
     /** Tries to maximize the objective value of the problem which is being solved */
     SearchStatistics maximize();
     /** @return the value of the best solution in this decision diagram if there is one */
-    Optional<Integer> bestValue();
+    Optional<Double> bestValue();
     /** @return the solution leading to the best solution in this decision diagram (if it exists) */
     Optional<Set<Decision>> bestSolution();
 }
